Address CR comments - Simplify DB seeding steps, rename validators/add_locations.js -> validators/locations.js
