+++ lib/plugins/features/movies/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Controller             = require('./controller');
-const AddLocationsValidator  = require('../../../validators/add_locations');
+const AddLocationsValidator  = require('../../../validators/location');
 const MovieValidator         = require('../../../validators/movie');
 const RetrieveQueryValidator = require('../../../validators/retrieve_query');
